requirements.txt
django_rest_framework
django-admin-row-actions
django-inline-actions
